Make logging to Telegram better.
Use forked telegram_hook that lets me change the log level on the fly and fixes formatting for errors.
Add a config value for what the telegram log level should be.
Use info level logging for telegram for the startup message (so that doesn't have to be an error) and then use that config value.