Do a better job polling for submissions. Breaks old databases.
Submission IDs should be numbers, not strings. This allows us to do a comparison against search results, so we don't think the entire page is new when the last submission we saw was deleted (or otherwise not present in the results).
Add a check to see if the user has seen the submission already before telling them about it.
Modify the search alert format string.
Notify users in separate goroutines (but the searches themselves are still synchronous).
HTML escape search strings and submission titles.